Fix broken .t files
All but one of these changed .t files set the locale to 'en_US' without
checking the return value for success.  Some platforms do not have this
locale, and so the setlocale() fails, leaving the locale unchanged from
the inherited value from the environment.  If that locale has a comma
for the radix character, for example, the tests fail.

format_bytes.t is different; it doesn't even bother trying to change the
locale, so will get whatever the inherited value is, and will succeed or
fail like the other tests.

This commit changes to use the C locale for the tests, as that should
always be available everywhere